// apps/storefront/src/generated/RouteBinding.ts
export const ROUTES = {
  storehome: '/',
  storebenefits: '/benefits',
  storecart: '/cart',
  storecheckout: '/checkout',
  storenotifications: '/notifications',
  storeorders: '/orders',
  storeorder: '/orders/:orderId',
  storeaftersale: '/orders/:orderId/aftersales',
  storepayment: '/payments/:paymentId/result',
  storecatalog: '/products',
  storeproduct: '/products/:productId',
  storeprofile: '/profile',
  storesecurity: '/profile/security',
  storesupport: '/support',
  storesupportcase: '/support/:caseId',
  storevouchers: '/vouchers',
} as const;